Remove glade from AccountPickerPopup
For some reason, the Widgets coming from the .glade file we were using
to lay out the AccountPickerPopup window were misbehaving.

As the use of glade here was part of the legacy of the original
AbstractWindow implementation, and not stictly necessary, reimplement
the Window instantiation directly in the AccountPickerPopup inner
class. Only added a few lines, and if anything the code looks better.

Not exactly a perfect resolution, but things are working now, which
is really the point here.